requirements.txt
asgiref==3.3.4
databases==0.4.3
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
